pr #36 review: address all human + coderabbit findings
Bug fixes:
  - TOM_TAB >> 6 always wrote zero — coreboot writes raw values
  - C0REFRCTRL+3 used setbits32 (unaligned 32-bit) — coreboot uses
    setbits8. Fixed to byte read-modify-write.
  - CAR window 0x100000 (1 MiB) exceeds D410's 512 KiB L2 — NEM
    can't pin more cache than exists. Fixed to 0x80000 + base adjusted.
  - _car_done missing ret — execution fell into .word MTRR data.
    Added ret before _fixed_mtrr_list.
  - MP flight_cpu_init infinite recursion — CPU_INIT_FN stored
    flight_cpu_init which then loaded and called itself. Fixed with
    monomorphized trampoline: make_cpu_init_trampoline<C: CpuOps>()
    returns a fn() that loads CPU_OPS_PTR and calls C::init_cpu().

Code organization (per reviewer):
  - New crate fstart-ecam (131 lines): global ECAM PCI config access
    extracted from fstart-pineview-regs. Not pineview-specific.
    pineview-regs re-exports for backward compat.
  - Removed EcamPci struct — redundant with ecam:: free functions.
  - ICH7 constants (LPC_DEV, SMBUS_DEV, etc.) moved to
    fstart-driver-intel-ich7::ich7 module. pineview-regs keeps
    backward-compat re-exports.
  - rdmsr/wrmsr moved to fstart-arch::msr module (x86_64 feature).
    Added Msr { lo, hi } struct with From<u64> / Into<u64>.
    cpu-pineview now imports from fstart-arch.
  - SPD split: ddr2.rs holds DDR2-specific constants and decode_dimm.
    lib.rs keeps shared types (DimmInfo, ChipWidth, etc.) + read_spd.

API improvements (per reviewer):
  - PineviewCpuOps: now takes pmbase from constructor instead of
    hardcoded PMB0_BASE constant. SB programs PMBASE, CPU shouldn't
    assume the value.
  - SmBus trait: added read_word, write_word, block_read, block_write
    with default impls returning HardwareError. I801SmBus wires
    read_word via existing read_word_data.
  - CK505: register arrays now [u8; 16] with num_regs field (was
    fixed [u8; 5]). Clock generators can have more or fewer than 5.
  - SuperIO: named constants for all common config registers
    (SIO_REG_LDN, SIO_REG_ENABLE, SIO_REG_IO_BASE_HI/LO,
    SIO_REG_IO_BASE2_HI/LO, SIO_REG_IRQ, SIO_REG_IRQ1,
    SIO_REG_DEVID_HI/LO). No more magic 0x30/0x60/0x70.
  - SuperIO: program_mouse uses SIO_REG_IRQ1 (0x72) for secondary
    IRQ. Documented that KBC+mouse can share one LDN — mouse IRQ
    goes in secondary register, no conflict.

Not changed (verified correct or deferred):
  - tock-registers for MchBar/Rcba: these have hundreds of registers
    at dynamically-computed offsets (gz(), gy(), etc.) which don't
    map to tock-registers' static register_structs! layout.
    The hostbridge/igd PCI config offsets are also not MMIO.
  - ICH7 CIR values: chipset-fixed initialization registers, not
    board-configurable. Same values across all ICH7 boards in
    coreboot.
  - LPC abstraction: keeping LpcBus for now — it's minimal (one
    method) and separates decode-range programming from PCI config.
    Future PCIe hotplug/bridges have different decode semantics.
  - CAS selection bitmask check: matches coreboot exactly — the
    loop checks SPD timing, not bitmask membership. Decrementing
    past unset bits is correct because the timing check fails for
    unsupported CAS anyway.
  - BSS clear before CAR: architectural issue for NEM boards where
    BSS is in CAR space. Requires platform-specific boot flow
    changes — tracked as future work.
