+++ Cargo.toml
@@ -35,6 +35,7 @@ members = [
   "crates/fstart-driver-fu740-prci",
   "crates/fstart-driver-fu740-ddr",
   "crates/fstart-driver-pci-ecam",
+  "crates/fstart-ecam",
   "crates/fstart-driver-bochs-display",
   "crates/fstart-device-registry",
   "crates/fstart-crabefi",
@@ -137,6 +138,7 @@ fstart-driver-sifive-uart = { path = "crates/fstart-driver-sifive-uart" }
 fstart-driver-fu740-prci = { path = "crates/fstart-driver-fu740-prci" }
 fstart-driver-fu740-ddr = { path = "crates/fstart-driver-fu740-ddr" }
 fstart-driver-pci-ecam = { path = "crates/fstart-driver-pci-ecam" }
+fstart-ecam = { path = "crates/fstart-ecam" }
 fstart-driver-bochs-display = { path = "crates/fstart-driver-bochs-display" }
 fstart-device-registry = { path = "crates/fstart-device-registry" }
 fstart-crabefi = { path = "crates/fstart-crabefi" }
