+++ crates/fstart-codegen/src/stage_gen/tests.rs
@@ -546,8 +546,9 @@ fn test_non_bus_parent_is_compile_error() {
     });
     parsed.driver_instances.push(DriverInstance::I2cCk505(
         fstart_device_registry::i2c_ck505::I2cCk505Config {
-            mask: [0, 0, 0, 0, 0],
-            regs: [0, 0, 0, 0, 0],
+            num_regs: 5,
+            mask: [0; 16],
+            regs: [0; 16],
         },
     ));
     // uart0 is at index 0
