+++ crates/fstart-cpu-pineview/src/lib.rs
@@ -11,6 +11,7 @@
 
 #![no_std]
 
+use fstart_arch::msr::{rdmsr, wrmsr};
 use fstart_mp::CpuOps;
 
 // ---------------------------------------------------------------------------
@@ -26,52 +27,8 @@ const MSR_PMG_IO_CAPTURE_ADDR: u32 = 0xE7;
 /// Miscellaneous feature enable.
 const IA32_MISC_ENABLE: u32 = 0x1A0;
 
-/// PM base I/O port (programmed by ICH7 southbridge).
-const PMB0_BASE: u32 = 0x0500;
-
 /// Maximum supported C-state level.
 const HIGHEST_CLEVEL: u32 = 3;
-
-// ---------------------------------------------------------------------------
-// MSR helpers
-// ---------------------------------------------------------------------------
-
-/// Read a 64-bit MSR.
-///
-/// # Safety
-///
-/// Caller must ensure `msr` is a valid MSR for this CPU.
-#[inline]
-unsafe fn rdmsr(msr: u32) -> u64 {
-    let lo: u32;
-    let hi: u32;
-    core::arch::asm!(
-        "rdmsr",
-        in("ecx") msr,
-        out("eax") lo,
-        out("edx") hi,
-        options(nomem, nostack),
-    );
-    ((hi as u64) << 32) | (lo as u64)
-}
-
-/// Write a 64-bit MSR.
-///
-/// # Safety
-///
-/// Caller must ensure `msr` and `val` are valid for this CPU.
-#[inline]
-unsafe fn wrmsr(msr: u32, val: u64) {
-    let lo = val as u32;
-    let hi = (val >> 32) as u32;
-    core::arch::asm!(
-        "wrmsr",
-        in("ecx") msr,
-        in("eax") lo,
-        in("edx") hi,
-        options(nomem, nostack),
-    );
-}
 
 // ---------------------------------------------------------------------------
 // Per-CPU configuration functions
@@ -84,7 +41,7 @@ unsafe fn wrmsr(msr: u32, val: u64) {
 /// I/O base/capture addresses.
 ///
 /// Matches coreboot `configure_c_states()` in model_106cx_init.c.
-fn configure_c_states() {
+fn configure_c_states(pmbase: u32) {
     // SAFETY: these MSRs are architecturally defined for Atom Pineview.
     unsafe {
         let mut cst = rdmsr(MSR_PKG_CST_CONFIG_CONTROL);
@@ -96,12 +53,12 @@ fn configure_c_states() {
 
         wrmsr(MSR_PKG_CST_CONFIG_CONTROL, cst);
 
-        // MWAIT IO base address (P_BLK).
-        let io_base = ((PMB0_BASE + 4) & 0xFFFF) as u64;
+        // MWAIT IO base address (P_BLK = PMBASE + 4).
+        let io_base = ((pmbase + 4) & 0xFFFF) as u64;
         wrmsr(MSR_PMG_IO_BASE_ADDR, io_base);
 
         // C-level controls: IO port + (highest_clevel - 2) in bits [18:16].
-        let io_capture = ((PMB0_BASE + 4) as u64) | (((HIGHEST_CLEVEL - 2) as u64) << 16);
+        let io_capture = ((pmbase + 4) as u64) | (((HIGHEST_CLEVEL - 2) as u64) << 16);
         wrmsr(MSR_PMG_IO_CAPTURE_ADDR, io_capture);
     }
 }
@@ -139,13 +96,23 @@ fn configure_misc() {
 ///
 /// Configures C-states, Enhanced SpeedStep, and thermal monitoring
 /// on every logical CPU during MP initialization.
-pub struct PineviewCpuOps;
+pub struct PineviewCpuOps {
+    /// PM base I/O port (programmed by the ICH7 southbridge).
+    pmbase: u32,
+}
+
+impl PineviewCpuOps {
+    /// Create with the southbridge's PM base I/O address.
+    pub fn new(pmbase: u32) -> Self {
+        Self { pmbase }
+    }
+}
 
 impl CpuOps for PineviewCpuOps {
     const NAME: &'static str = "Intel Atom Pineview (106cx)";
 
     fn init_cpu(&self) {
-        configure_c_states();
+        configure_c_states(self.pmbase);
         configure_misc();
         fstart_log::info!("cpu: Pineview MSR configuration complete");
     }
