+++ crates/fstart-driver-intel-ich7/Cargo.toml
@@ -12,6 +12,7 @@ acpi = ["dep:fstart-acpi", "dep:fstart-acpi-macros"]
 [dependencies]
 fstart-acpi = { workspace = true, features = ["x86"], optional = true }
 fstart-acpi-macros = { path = "../fstart-acpi-macros", optional = true }
+fstart-ecam = { workspace = true }
 fstart-gpio-ich = { workspace = true }
 fstart-hda = { workspace = true }
 heapless = { workspace = true }
